pyproject.toml
[tool.poetry]
name = "research-reader"
version = "0.1.0"
description = ""
authors = ["Example Author <example@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
ruff = "0.4.7"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.6"
#langchain = "0.2.5"
langchain-openai = "^0.1.10"
opencv-python = "4.9.0.80"
streamlit = "1.35.0"
chromadb = "0.5.0"
rank-bm25 = "0.2.2"
pypdf = "4.2.0"
aiohttp = "3.9.5"
aiosignal = "1.3.1"
altair = "5.3.0"
annotated-types = "0.7.0"
anyio = "4.4.0"
asgiref = "3.8.1"
attrs = "23.2.0"
backoff = "2.2.1"
bcrypt = "4.1.3"
beautifulsoup4 = "4.12.3"
blinker = "1.8.2"
build = "1.2.1"
cachetools = "5.3.3"
certifi = "2024.6.2"
charset-normalizer = "3.3.2"
chroma-hnswlib = "0.7.3"
click = "8.1.7"
coloredlogs = "15.0.1"
contourpy = "1.2.1"
cycler = "0.12.1"
dataclasses-json = "0.6.6"
deprecated = "1.2.14"
distro = "1.9.0"
dnspython = "2.6.1"
email-validator = "2.1.1"
fastapi = "0.111.0"
fastapi-cli = "0.0.4"
filelock = "3.14.0"
flatbuffers = "24.3.25"
fonttools = "4.53.0"
frozenlist = "1.4.1"
fsspec = "2024.5.0"
gitdb = "4.0.11"
gitpython = "3.1.43"
google-auth = "2.29.0"
googleapis-common-protos = "1.63.0"
greenlet = "3.0.3"
grpcio = "1.64.0"
h11 = "0.14.0"
httpcore = "1.0.5"
httptools = "0.6.1"
httpx = "0.27.0"
huggingface-hub = "0.23.2"
humanfriendly = "10.0"
idna = "3.7"
importlib-metadata = "7.1.0"
importlib-resources = "6.4.0"
jinja2 = "3.1.4"
jsonpatch = "1.33"
jsonpointer = "2.4"
jsonschema = "4.22.0"
jsonschema-specifications = "2023.12.1"
kiwisolver = "1.4.5"
kubernetes = "29.0.0"
langchain-core = "^0.2.10"
langchain-text-splitters = "0.2.0"
langsmith = "0.1.81"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
marshmallow = "3.21.2"
matplotlib = "3.9.0"
mdurl = "0.1.2"
mmh3 = "4.1.0"
monotonic = "1.6"
mpmath = "1.3.0"
multidict = "6.0.5"
mypy-extensions = "1.0.0"
numpy = "1.26.4"
oauthlib = "3.2.2"
onnxruntime = "1.18.0"
openai = "1.30.5"
opentelemetry-api = "1.25.0"
opentelemetry-exporter-otlp-proto-common = "1.25.0"
opentelemetry-exporter-otlp-proto-grpc = "1.25.0"
opentelemetry-instrumentation = "0.46b0"
opentelemetry-instrumentation-asgi = "0.46b0"
opentelemetry-instrumentation-fastapi = "0.46b0"
opentelemetry-proto = "1.25.0"
opentelemetry-sdk = "1.25.0"
opentelemetry-semantic-conventions = "0.46b0"
opentelemetry-util-http = "0.46b0"
orjson = "3.10.3"
overrides = "7.7.0"
packaging = "23.2"
pandas = "2.2.2"
pillow = "10.3.0"
posthog = "3.5.0"
protobuf = "4.25.3"
pyarrow = "16.1.0"
pyasn1 = "0.6.0"
pyasn1-modules = "0.4.0"
pydeck = "0.9.1"
pygments = "2.18.0"
pyparsing = "3.1.2"
pypika = "0.48.9"
pyproject-hooks = "1.1.0"
python-dateutil = "2.9.0.post0"
python-multipart = "0.0.9"
pytz = "2024.1"
pyyaml = "6.0.1"
referencing = "0.35.1"
regex = "2024.5.15"
requests = "2.32.3"
requests-oauthlib = "2.0.0"
rich = "13.7.1"
rpds-py = "0.18.1"
rsa = "4.9"
setuptools = "70.0.0"
shellingham = "1.5.4"
six = "1.16.0"
smmap = "5.0.1"
sniffio = "1.3.1"
soupsieve = "2.5"
sqlalchemy = "2.0.30"
starlette = "0.37.2"
sympy = "1.12.1"
tenacity = "8.3.0"
tiktoken = "0.7.0"
tokenizers = "0.19.1"
toml = "0.10.2"
toolz = "0.12.1"
tornado = "6.4"
tqdm = "4.66.4"
typer = "0.12.3"
typing-inspect = "0.9.0"
typing-extensions = "4.12.1"
tzdata = "2024.1"
ujson = "5.10.0"
urllib3 = "2.2.1"
uvicorn = "0.30.0"
uvloop = "0.19.0"
watchdog = "4.0.1"
watchfiles = "0.22.0"
websocket-client = "1.8.0"
websockets = "12.0"
wheel = "0.43.0"
wrapt = "1.16.0"
yarl = "1.9.4"
zipp = "3.19.1"
cryptography = "^3.1"
langchain = "^0.2.6"
unstructured = {extras = ["all-docs"], version = "^0.14.9"}
psutil = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
